+++ zapx/truncated_string_test.go
@@ -66,12 +66,10 @@ func TestTruncatedStringField_String(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			field := truncatedStringField{
-				value:     tt.value,
+				value:     stringer(tt.value),
 				maxLength: tt.maxLength,
 			}
 			require.Equal(t, tt.want, field.String())
 		})
 	}
 }
-
-